fix problem exception atmosphere
